refactor: CreateTaskPage
Separate hooks and types into separate files.
