+++ src/domains/tasks/CreateTask/index.tsx
@@ -1,1 +1,2 @@
-export * from "./components";
+export { CreateTaskPage } from "./components";
+export * from "./types";
