+++ src/v2/components/index.tsx
@@ -2,3 +2,4 @@ export * from "./LayoutV2";
 export * from "./Navigation";
 export * from "./TopBar";
 export * from "./Select";
+export * from "./TextField";
